+++ lib/kward/cli/runtime_helpers.rb
@@ -94,7 +94,7 @@ module Kward
           return true
         end
 
-        expanded_command = bang_shell(agent).expand_alias(command)
+        expanded_command = bang_shell(agent).expand_alias(command, interactive: true)
         run_user_interactive_pty_command(
           expanded_command,
           shell: Ekwsh::DEFAULT_SHELL,